Tests, added permission check for API
