+++ python/aristotle-metadata-registry/aristotle_mdr/contrib/publishing/models.py
@@ -57,17 +57,10 @@ class VersionPermissions(TimeStampedModel):
     def __str__(self):
         return "Version is: {}  and permissions are: {}".format(str(self.version), str(self.visibility))
 
-    def can_view(self):
-        return True
-        # TODO implement API permissions
-
-    def can_edit(self):
-        return True
-        # TODO implement API permissions
-
     @property
     def id(self):
         return self.version.id
+
 
 
 class PublicationRecord(TimeStampedModel):
